Update and rename TemplateMod.cs to UnlimitedPolicies.cs
Simplified code a lot, making it a bit more robust, definitely faster.

To build your mod you are going to need to grab the PhotonUnityNetworking.dll library from the game's files and add it to the Libraries folder of your mod, and register it in your .csproj file as well

This will clear the error coming from MultiplayerManager that depends on that lib

Other than that, you're good, just tested it, it works!
